code-review-agent/requirements.txt
# Core
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# LLM Providers
openai>=1.50.0
ollama>=0.3.0

# Web Framework
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12

# GitHub Integration
PyGithub>=2.4.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx>=0.27.0
aiohttp>=3.10.0

# Code Analysis
tree-sitter>=0.23.0
pygments>=2.18.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# Development
black>=24.0.0
ruff>=0.6.0
